origin -> cpython
I don't know how in the world someone was running this
previously, but now it'll work with adding a remote called
cpython.

Probably it's easier to not do this this way anyhow and
to instead just directly hit GitHub for those files, but
whatever, for now this works.

And this needs adding to the README I guess.